chore: update VotingPowerProvider
