+++ apps/evm/src/lib/bar/index.ts
@@ -1,4 +1,2 @@
 export * from './useBarChartData'
 export * from './useBarData'
-export * from './useVotingPower'
-export * from './useVotingPowerBreakdown'
