+++ apps/evm/src/ui/bar/VotingPowerProvider.tsx
@@ -1,13 +1,55 @@
 'use client'
 
-import { useAccount } from '@sushiswap/wagmi'
-import { FC, ReactNode, createContext, useContext, useMemo } from 'react'
+import snapshot from '@snapshot-labs/snapshot.js'
 import {
-  useVotingPower as useVotingPowerQuery,
-  useVotingPowerBreakdown,
-} from 'src/lib/bar'
-import { Amount, Type } from 'sushi/currency'
+  MASTERCHEF_ADDRESS,
+  useAccount,
+  useBalancesWeb3,
+  useContractReads,
+} from '@sushiswap/wagmi'
+import { useQuery } from '@tanstack/react-query'
+import { FC, ReactNode, createContext, useContext, useMemo } from 'react'
+import { erc20Abi, masterChefV1Abi } from 'sushi/abi'
+import { ChainId } from 'sushi/chain'
+import {
+  Amount,
+  SUSHI_ADDRESS,
+  Token,
+  Type,
+  XSUSHI,
+  XSUSHI_ADDRESS,
+  tryParseAmount,
+} from 'sushi/currency'
 import { Fraction } from 'sushi/math'
+
+const SnapshotStrategies = [
+  {
+    name: 'erc20-balance-of',
+    params: {
+      address: '0x62d11bc0652e9D9B66ac0a4c419950eEb9cFadA6',
+      symbol: 'SUSHIPOWAH',
+      decimals: '18',
+    },
+  },
+  {
+    name: 'erc20-balance-of',
+    params: {
+      address: '0x6811079E3c63ED96Eb005384d7E7ec8810E3D521',
+      decimals: 18,
+    },
+    network: '137',
+  },
+]
+
+const SUSHI_ETH_SLP_ADDRESS = '0x795065dCc9f64b5614C407a6EFDC400DA6221FB0'
+
+export const SUSHI_ETH_SLP = new Token({
+  chainId: ChainId.ETHEREUM,
+  address: SUSHI_ETH_SLP_ADDRESS,
+  decimals: 18,
+  symbol: 'SUSHI-ETH SLP',
+  name: 'SUSHI-ETH LP Token',
+})
 
 interface VotingPowerContext {
   votingPower: number | undefined
@@ -39,25 +81,179 @@ export const VotingPowerProvider: FC<{
   const { address, isConnected } = useAccount()
 
   const {
-    data: votingPower,
-    isLoading,
-    isError,
-  } = useVotingPowerQuery({ address })
+    data: votingPowerData,
+    isLoading: isVotingPowerLoading,
+    isError: isVotingPowerError,
+  } = useQuery({
+    queryKey: ['snapshot-scores', address],
+    queryFn: async () => {
+      if (!address) throw new Error()
 
-  const { balances, weights } = useVotingPowerBreakdown({ address })
+      const resp = await snapshot.utils.getVp(
+        address,
+        '1',
+        SnapshotStrategies,
+        'latest',
+        'sushigov.eth',
+        true,
+      )
+
+      return resp
+    },
+    enabled: Boolean(address),
+    staleTime: 300000,
+  })
+
+  const {
+    data: ethereumBalances,
+    isLoading: isBalancesLoading,
+    isError: isBalancesError,
+  } = useBalancesWeb3({
+    chainId: ChainId.ETHEREUM,
+    currencies: [XSUSHI[ChainId.ETHEREUM], SUSHI_ETH_SLP],
+    account: address,
+    enabled: Boolean(address),
+  })
+
+  const {
+    data: contractData,
+    isLoading: isContractDataLoading,
+    isError: isContractDataError,
+  } = useContractReads({
+    contracts: useMemo(
+      () => [
+        // sushi.balanceOf(SUSHI-ETH LP)
+        {
+          chainId: ChainId.ETHEREUM,
+          abi: erc20Abi,
+          address: SUSHI_ADDRESS[ChainId.ETHEREUM],
+          functionName: 'balanceOf',
+          args: [SUSHI_ETH_SLP_ADDRESS],
+        },
+        // sushi.balanceOf(XSUSHI)
+        {
+          chainId: ChainId.ETHEREUM,
+          abi: erc20Abi,
+          address: SUSHI_ADDRESS[ChainId.ETHEREUM],
+          functionName: 'balanceOf',
+          args: [XSUSHI_ADDRESS[ChainId.ETHEREUM]],
+        },
+        // SUSHI-ETH-LP.totalSupply()
+        {
+          chainId: ChainId.ETHEREUM,
+          abi: erc20Abi,
+          address: SUSHI_ETH_SLP_ADDRESS,
+          functionName: 'totalSupply',
+        },
+        // xsushi.totalSupply()
+        {
+          chainId: ChainId.ETHEREUM,
+          abi: erc20Abi,
+          address: XSUSHI_ADDRESS[ChainId.ETHEREUM],
+          functionName: 'totalSupply',
+        },
+        // user's staked SUSHI-ETH LP
+        ...(address
+          ? [
+              {
+                chainId: ChainId.ETHEREUM,
+                abi: masterChefV1Abi,
+                address: MASTERCHEF_ADDRESS[ChainId.ETHEREUM],
+                functionName: 'userInfo',
+                args: [12n, address],
+              },
+            ]
+          : []),
+      ],
+      [address],
+    ),
+    staleTime: 300000,
+  })
+
+  const weights = useMemo(() => {
+    if (
+      !contractData?.length ||
+      contractData.some((data) => data.status !== 'success')
+    )
+      return undefined
+
+    const [
+      sushiBalanceSLP,
+      sushiBalanceXSUSHI,
+      slpTotalSupply,
+      xsushiTotalSupply,
+    ] = contractData
+
+    const xsushiWeight = new Fraction(
+      sushiBalanceXSUSHI.result as bigint,
+      xsushiTotalSupply.result as bigint,
+    )
+
+    const slpWeight = new Fraction(
+      (sushiBalanceSLP.result as bigint) * 2n,
+      slpTotalSupply.result as bigint,
+    )
+
+    return {
+      xsushi: xsushiWeight,
+      slp: slpWeight,
+      xsushiPolygon: new Fraction(1, 1),
+    }
+  }, [contractData])
+
+  const balances = useMemo(() => {
+    if (
+      !ethereumBalances ||
+      !votingPowerData?.vp_by_strategy?.length ||
+      !contractData?.[4]?.result
+    ) {
+      return undefined
+    }
+
+    const userStakedSLP = contractData[4].result as [bigint, bigint]
+
+    return {
+      xsushi: ethereumBalances[XSUSHI_ADDRESS[ChainId.ETHEREUM]],
+      slp: ethereumBalances[SUSHI_ETH_SLP_ADDRESS].add(
+        Amount.fromRawAmount(SUSHI_ETH_SLP, userStakedSLP[0]),
+      ) as Amount<Type>,
+      xsushiPolygon: tryParseAmount(
+        votingPowerData.vp_by_strategy[1].toString(),
+        new Token({
+          chainId: ChainId.POLYGON,
+          address: XSUSHI_ADDRESS[ChainId.POLYGON],
+          decimals: 18,
+          symbol: 'XSUSHI (Polygon)',
+          name: 'SushiBar',
+        }),
+      ) as Amount<Type>,
+    }
+  }, [contractData, ethereumBalances, votingPowerData])
 
   return (
     <Context.Provider
       value={useMemo(
         () => ({
-          votingPower,
+          votingPower: votingPowerData?.vp,
           balances,
           weights,
           isConnected,
-          isLoading,
-          isError,
+          isLoading:
+            isVotingPowerLoading || isBalancesLoading || isContractDataLoading,
+          isError: isVotingPowerError || isBalancesError || isContractDataError,
         }),
-        [votingPower, balances, weights, isConnected, isError, isLoading],
+        [
+          votingPowerData,
+          balances,
+          weights,
+          isConnected,
+          isVotingPowerLoading,
+          isBalancesLoading,
+          isContractDataLoading,
+          isVotingPowerError,
+          isBalancesError,
+          isContractDataError,
+        ],
       )}
     >
       {children}
